requirements.txt
flask
requests
pandas
tensorflow
scikit-learn
h5py
python-dotenv
numpy
langchain
openai
streamlit
ast
chains
langchain-community
tiktoken 
faiss-cpu
langchain-openai
streamlit_chat
sentence-transformers
unstructured
langchain-huggingface
mlflow
joblib
guardrails-ai
torch
torchvision
seaborn